requirements.txt
requests==2.11.1
Flask==1.0.2
PyYAML==5.1
RPi.GPIO